✌ A dead-simple Gui system. I was massively inspired to create this after I spent a day modifying PlayerParticles. As such, this system works very similarly but is dead simple to use when developing and honestly just makes me happy to use.
This is going to be a major improvement over my previous Gui systems (check Phase for example, it's kind of a mess- each Gui had a GuiType, dedicated handler class, a Gui menu class, and a registery in the click listener. Not pretty).

So yeah, safe to say I'm a bit excited to use this >:)