new ConcurrentUtils#callAsync flavor 😌
this new version removes the need for returning null every time at the end of the callable by using a custom consumer that throws exception.

i also want to replace the original callAsync method with a more generic one that will allow returning object types, but that'll have to be wait until this method is fully phased out since the method signatures will clash.